sprint-123: T-12301 Extract ferric-cli into a library with thin binary shims
