+++ docs/intents/INT-0009-lean-decomposed-architecture.md
@@ -1,6 +1,6 @@
 # INT-0009 — Lean, decomposed harness architecture
 
-- **State:** planned
+- **State:** active
 - **Work evidence:** [Sprint 123 ferric-cli library extraction plan](../sprints/s123/sprint-plans/build-plan.md#execution-sequence); [direction & refactor plan](../plans/2026-09-06-direction-and-refactor.md)
 - **Test evidence:** (none yet)
 
@@ -75,3 +75,5 @@ because the code stays in one crate. Later increments (serving-layer crate,
 - 2026-09-06: moved `proposed` → `planned` when Sprint 123's Plan locked the
   library-extraction increment (T-12301), targeting AC-1; AC-2/AC-4 remain
   proposed for later increments.
+- 2026-09-06: moved `planned` → `active` when Sprint 123's Build began the
+  library extraction (T-12301).
