+++ docs/work/tasks.md
@@ -308,3 +308,5 @@ below remain executable follow-up work; the review did not silently fix them.
 
 - [ ] T-12204 (backlog) [intent: INT-0008]: Add a front-door test seam — a `#[cfg(test)]` `Startup` constructor accepting a synthetic model list, or an injectable `MemoryProbe` threaded into `session_with` — so the won't-fit picker gate gains a full driver E2E rather than only decision-helper coverage (Sprint 122 C-001). — touches: `crates/ferric-cli/src/startup.rs`, `crates/ferric-cli/src/human.rs`, front-door tests
 - [ ] T-12205 (backlog, after T-12204) [intent: INT-0008]: Extend hardware fit to GPU/VRAM discovery and safe layer calibration (the GPU half of T-11507), and feed both RAM and VRAM fit into AC-13's acquisition recommendation. — touches: `crates/ferric-cli/src/startup/memory.rs`, `crates/ferric-cli/src/backend.rs`, `crates/ferric-cli/src/server.rs`
+
+## Sprint 123 — ferric-cli library extraction (INT-0009 AC-1)
